requirements.txt
pytesseract
pillow
opencv-python
numpy
python-dotenv
google-genai